Refactor routers This refactors the routes file into separate routers that contain grouped logic, like the content router.
It uses a better naming convention for controllers so that it's clear
what the file is when just looking at file names.
