+++ app/controllers/elements-controller.js
@@ -1,0 +1,7 @@
+function index(req, res) {
+  res.render('elements');
+}
+
+module.exports = {
+  index,
+};
