+++ config/routes/index.js
@@ -1,0 +1,9 @@
+const contentRouter = require('./content-router');
+const developmentRouter = require('./development-router');
+const appStatusRouter = require('./app-status-router');
+
+module.exports = [
+  appStatusRouter,
+  contentRouter,
+  developmentRouter,
+];
